Albert Chin found another signed/unsigned problem in the date and time
* xmlschemastypes.c: Albert Chin found another signed/unsigned problem
  in the date and time code raised on IRIX 6.5
Daniel

+++ xmlschemastypes.c
@@ -3759,8 +3759,8 @@ _xmlSchemaDateAdd (xmlSchemaValPtr dt, xmlSchemaValPtr dur)
 
     while (1) {
         if (tempdays < 1) {
-            long tmon = (long) MODULO_RANGE(r->mon-1, 1, 13);
-            long tyr  = r->year + (long)FQUOTIENT_RANGE(r->mon-1, 1, 13);
+            long tmon = (long) MODULO_RANGE((int)r->mon-1, 1, 13);
+            long tyr  = r->year + (long)FQUOTIENT_RANGE((int)r->mon-1, 1, 13);
             if (tyr == 0)
                 tyr--;
             tempdays += MAX_DAYINMONTH(tyr, tmon);
